Split behavior parent default tests
