+++ docs/COMPLETION_AUDIT.md
@@ -2863,6 +2863,11 @@ and do not assume Phase 4 is ready without evidence.
   preserving variant payload count/type, visibility, typed payload,
   generic enum payload, variant-list, and owner diagnostics while keeping
   struct metadata validation focused.
+- Resolver-restored behavior parent default synthesis tests now live in
+  `src/typechecker/tests/resolver_collection/behavior_parents/default_synthesis.rs`,
+  preserving inherited default method synthesis and generic default return
+  substitution coverage while keeping parent restoration, duplicate, conflict,
+  and cycle diagnostics focused.
 - Effect checking, typed allocator semantics, actors in std integration,
   JSON/YAML IR boundaries, and broader build graph execution remain gated by
   `docs/V1_SPEC.md`.
